model saved as .joblibs in trained_model
