+++ AI_Model_Forest/requirements.txt
@@ -3,3 +3,4 @@ numpy
 scikit-learn
 astropy
 joblib
+imbalanced-learn
