+++ AI_Model_Forest/trained_model/feature_columns.txt
@@ -9,3 +9,6 @@ STELLAR_LOGG
 STELLAR_RADIUS
 RA_DECIMAL
 DEC_DECIMAL
+R_PLANET_R_STAR_RATIO
+DEPTH_PER_RADIUS
+STELLAR_DENSITY_PROXY
